+++ xoeuf/models/proxy.py
@@ -1,16 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # ---------------------------------------------------------------------
-# xoeuf.models
-# ---------------------------------------------------------------------
-# Copyright (c) 2016-2017 Merchise Autrement [~º/~] and Contributors
+# Copyright (c) Merchise Autrement [~º/~] and Contributors
 # All rights reserved.
 #
-# This is free software; you can redistribute it and/or modify it under the
-# terms of the LICENCE attached (see LICENCE file) in the distribution
-# package.
+# This is free software; you can do what the LICENCE file allows you to.
 #
-# Created on 2016-08-24
 
 '''Transparently import models.
 
@@ -71,6 +66,7 @@ class ModelImporter(object):
         res = sys.modules.get(fullname, None)
         if res is None:
             res = ModelProxy(fullname.rsplit(str('.'), 1)[-1])
+            res.__path__ = __path__
             sys.modules[fullname] = res
         return res
 
@@ -97,6 +93,7 @@ class ModelImporter(object):
 
 class ModelProxy(ModuleType, _proxy.ModelProxy):
     def __init__(self, name):
+        ModuleType.__init__(self, name)
         _proxy.ModelProxy.__init__(self, name)