+++ xoeuf/tests/test_enumeration/tests/test_enumeration.py
@@ -1,20 +1,17 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # ---------------------------------------------------------------------
-# test_property
-# ---------------------------------------------------------------------
-# Copyright (c) 2017 Merchise Autrement [~º/~] and Contributors
+# Copyright (c) Merchise Autrement [~º/~] and Contributors
 # All rights reserved.
 #
-# This is free software; you can redistribute it and/or modify it under the
-# terms of the LICENCE attached (see LICENCE file) in the distribution
-# package.
+# This is free software; you can do what the LICENCE file allows you to.
 #
-# Created on 2017-08-01
 
 from __future__ import (division as _py3_division,
                         print_function as _py3_print,
                         absolute_import as _py3_abs_import)
+
+import contextlib
 
 from hypothesis import strategies as s, given
 from xoeuf.odoo.tests.common import TransactionCase
@@ -54,20 +51,21 @@ class TestEnum(TransactionCase):
         assert isinstance(obj.color, COLORS)
 
     def test_cannot_set_invalid_integers(self):
-        # Unfortunately we can insert invalid values in the DB, but we get a
-        # ValueError upon 'reading'.
-        obj = self.EnumModel.create({'color': 10})
-        id = obj.id
-        self.EnumModel.invalidate_cache()
-        obj = self.EnumModel.browse(id)
-        with self.assertRaises(ValueError):
-            obj.color
+        # Sinces tests are run while the registry is being populated, i.e not
+        # ready, we need to trick it to allow receivers be executed.
+        with force_ready(self.env.registry), self.assertRaises(ValueError):
+            self.EnumModel.create({'color': 10})
 
-    def test_cannot_set_invalid_integers2(self):
-        # Unfortunately we can insert invalid values in the DB, but we get a
-        # ValueError upon 'reading'.
-        obj = self.EnumModel.create({'color': 10})
-        id = obj.id
-        self.EnumModel.invalidate_cache()
-        with self.assertRaises(ValueError):
-            self.EnumModel.browse(id).read(['color'])
+    def test_cannot_write_invalid_integers(self):
+        obj = self.EnumModel.create({'color': 1})
+        with force_ready(self.env.registry), self.assertRaises(ValueError):
+            obj.write({'color': 10})
+
+
+@contextlib.contextmanager
+def force_ready(registry):
+    registry.ready = True
+    try:
+        yield
+    finally:
+        registry.ready = False