+++ xoeuf/tests/test_localized_dt/__openerp__.py
@@ -1,21 +1,16 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # ---------------------------------------------------------------------
-# __openerp__
-# ---------------------------------------------------------------------
-# Copyright (c) 2017 Merchise Autrement [~º/~] and Contributors
+# Copyright (c) Merchise Autrement [~º/~] and Contributors
 # All rights reserved.
 #
-# This is free software; you can redistribute it and/or modify it under the
-# terms of the LICENCE attached (see LICENCE file) in the distribution
-# package.
+# This is free software; you can do what the LICENCE file allows you to.
 #
-# Created on 2017-03-07
 
-dict(
-    name='test_localized_dt',
-    version='1.0',
-    depends=['base', ],
-    auto_install=False,
-    installable=ODOO_VERSION_INFO[0] in (8, 9, 10)   # noqa
-)
+{
+    'name': 'test_localized_dt',
+    'version': '1.0',
+    'depends': ['base', ],
+    'auto_install': False,
+    'installable': True
+}