+++ xoeuf/tests/test_proxy/__openerp__.py
@@ -1,22 +1,17 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # ---------------------------------------------------------------------
-# __openerp__
-# ---------------------------------------------------------------------
-# Copyright (c) 2017 Merchise Autrement [~º/~] and Contributors
+# Copyright (c) Merchise Autrement [~º/~] and Contributors
 # All rights reserved.
 #
-# This is free software; you can redistribute it and/or modify it under the
-# terms of the LICENCE attached (see LICENCE file) in the distribution
-# package.
+# This is free software; you can do what the LICENCE file allows you to.
 #
-# Created on 2017-10-03
 
-dict(
-    name='test_proxy',
-    author='Merchise Autrement [~º/~] and Contributors',
-    description='Test for the xoeuf.models.proxy',
-    depends=['base', 'web', ],
-    installable=True,
-    auto_install= False,
-)
+{
+    'name': 'test_proxy',
+    'author': 'Merchise Autrement [~º/~] and Contributors',
+    'description': 'Test for the xoeuf.models.proxy',
+    'depends': ['base', 'web', ],
+    'installable': True,
+    'auto_install':  False,
+}