+++ xoeuf/tests/test_proxy/tests/test_proxy.py
@@ -11,6 +11,12 @@ from __future__ import (division as _py3_division,
                         print_function as _py3_print,
                         absolute_import as _py3_abs_import)
 
+try:
+    import unittest2 as unittest
+except ImportError:
+    import unittest
+
+from xoeuf import MAJOR_ODOO_VERSION
 from xoeuf.models.proxy import ResUsers as Users
 from xoeuf.odoo.tests.common import TransactionCase, HttpCase
 
@@ -23,23 +29,34 @@ class TestModelProxy(TransactionCase):
     def test_finds_self(this):
         self = this.env['ir.model']  # noqa: make self a valid recordset
         Users.create({'name': 'John Doe', 'login': 'john.doe'})
-        Users.search([])
+        this.assertEqual(Users.search([]), self.env['res.users'].search([]))
 
 
 class TestHTTPModelProxy(HttpCase):
+    at_install = False
+    post_install = not at_install
+
+    @staticmethod
+    def getcode(response):
+        if MAJOR_ODOO_VERSION < 11:
+            return response.getcode()
+        else:
+            # Odoo 11+ uses requests to load the URL.
+            return response.status_code
+
+    @unittest.skipIf(MAJOR_ODOO_VERSION >= 9, 'Does not raises the error in Odoo 9+')
     def test_request_no_auth(self):
         # The controller won't be able to find a proper environment and fail
         # with an AttributeError, we'll see as an error 500
         self.authenticate('admin', 'admin')
         response = self.url_open('/test_proxy_none')
-        code = response.getcode()
-        assert code == HTTP_SERVER_ERROR, \
-            'We expected a server error, got %r.' % code
+        code = self.getcode(response)
+        self.assertEqual(code, HTTP_SERVER_ERROR)
 
     def test_request_with_auth(self):
         response = self.url_open('/test_proxy_pub')
-        code = response.getcode()
-        assert code == HTTP_OK, 'We expected an OK response, got %r.' % code
+        code = self.getcode(response)
+        self.assertEqual(code, HTTP_OK)
 
 
 HTTP_OK = 200